feat: add fault injection middleware to the server (#9)
This change adds a middleware around all routes in the test server that
can be used to trigger faults when serving requests. The middleware is
only enabled on requests with `Request-ID` AND `Fault-Settings` headers
set. The `Fault-Settings` header is a JSON object which is used to
configure whether or not to delay responses, reject requests and/or
return HTTP error codes.