fix: Remove GET/HEAD method request body handling
Reference: https://github.com/whatwg/fetch/issues/551

Certain interpretations and implementations of the IETF RFC do not support GET/HEAD HTTP method request body content. The request body checking is extraneous for the purpose of this testing, which is only intended for just the HTTP methods themselves, so just removing it.