feat: add oauth2 and ecommerce test apis
\## OAuth2

This change adds a new OAuth2 middleware and token endpoint that can be
used to enforce OAuth2 on any test endpoints. Test code can opt into
OAuth2 enforcement by setting the `x-require-oauth2: 1` header on any
SDK methods. This is usually done with a custom http client, per-method
custom headers or a SDK-wide hook depending on the language and test.

The following OAuth2 flows are supported:

- Client Credentials
- Authorization Code
- Resource Owner Password Credentials
- Refresh Token

When requesting an access token against `POST /oauth2/token`, the
following credentials are expected based on the flow used:

- Client ID: `beezy`
- Client Secret: `super-secret`
- Username: `testuser`
- Password: `testpassword`
- Code: `secret-auth-code`
- Refresh Token: `secret-refresh-token`

Test code can look for the `x-oauth2: pass` header to determine if the
test service detected and validated OAuth2 access tokens in the response
of any non-auth endpoints

\## E-commerce API

This is a new set of test APIs that are inspired by real world
e-commerce APIs. These APIs can enforce OAuth2 scopes if incoming
requests have passed the OAuth2 middleware.